Delete ProjectHome.md
Project is on hold since I don't have time for it currently and the work for TE was not as much fun as I expected. Without any feedback and time to rewrite the core I'll just remove the "Current State" message and leave it as some sort of testing ground.